[ UPDATE-login ] 유저가 로그인 한 경우 navbar의 Text를 변경하라 - localStorage에 있는 데이터로 navbar의 Text를 변경하였으나, 바로 업데이트 되지 않는 버그 발생 ( 로그인, 로그아웃 시) - contextAPI를 사용하여 로그인 한 유저의 정보를 가져와 사용하였다. - 유저의 로그인을 유지하는 fetchUser함수를 AuthContext로 가져왔다. - 페이지가 렌더링 될 때마다 me를 요청하고 jwt를 확인하여 로그인 유저임을 인지하게 하였고, - 로그인 페이지에서만 했던 부분을 전체로 확대하여 nav, 기타 페이지에서 사용할 예정이다.
